fix: cast all statvfs fields to u64 for macOS compatibility
On macOS all statvfs fields (f_blocks, f_bfree, f_bavail,
f_frsize, f_bsize) are u32, while on Linux they are u64.
Cast every field to u64 explicitly with #[allow(unnecessary_cast)]
so the same code compiles and works on both platforms.

+++ src/fs_utils.rs
@@ -88,12 +88,16 @@ pub fn disk_stats(path: &Path) -> (u64, u64, u64) {
             return (0, 0, 0);
         }
         let st = unsafe { st.assume_init() };
-        // f_frsize/f_bsize are u32 on macOS, u64 on Linux — cast is needed on macOS
+        // All statvfs fields are u32 on macOS and u64 on Linux.
+        // Cast everything to u64 explicitly; the cast is a no-op on Linux.
         #[allow(clippy::unnecessary_cast)]
         let block_size = if st.f_frsize > 0 { st.f_frsize as u64 } else { st.f_bsize as u64 };
-        let total = st.f_blocks.saturating_mul(block_size);
-        let used  = st.f_blocks.saturating_sub(st.f_bfree).saturating_mul(block_size);
-        let avail = st.f_bavail.saturating_mul(block_size);
+        #[allow(clippy::unnecessary_cast)]
+        let total = (st.f_blocks as u64).saturating_mul(block_size);
+        #[allow(clippy::unnecessary_cast)]
+        let used  = (st.f_blocks as u64).saturating_sub(st.f_bfree as u64).saturating_mul(block_size);
+        #[allow(clippy::unnecessary_cast)]
+        let avail = (st.f_bavail as u64).saturating_mul(block_size);
         (total, used, avail)
     }
     #[cfg(not(unix))]
